Format string after judge level in pglog
This commit change the behavior of the gplog package when record
debug and verbose level logs.

When the log level set to "Info", a "Debug" or "Verbose" function
call should do less job. The old code will a format for the log
message before judge the log level, buf if the log level is "Info",
the format call is an redundant operation.

So, we change it to judge the log level before format the
format operation.