Add Custom command for more logging control
The Progress command was just for doing info level to the logfile and verbose
level to stdout for table progress in gpbackup and gprestore. However, I found
at least one more place where using different settings for terminal and logfile
would be useful. We CheckClusterError in the cluster package for some commands
and assume we will log details to the file, but not to stdout.

Since this is a more general use case, we can make a more general command that
lets the caller pass in values for file threshold and shell threshold
separately to accommodate more use cases.

Keep Progress command as a wrapper around the Custom command since it is a
specific instance of the Custom usage